refactor

+++ src/daft-core/src/python/series.rs
@@ -268,7 +268,7 @@ impl PySeries {
         Ok(self.series.utf8_split(&pattern.series)?.into())
     }
 
-    pub fn utf8_extract(&self, pattern: &Self, index: i32) -> PyResult<Self> {
+    pub fn utf8_extract(&self, pattern: &Self, index: usize) -> PyResult<Self> {
         Ok(self.series.utf8_extract(&pattern.series, index)?.into())
     }
